+++ sbercloud/provider.go
@@ -4,14 +4,17 @@ import (
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/cts"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/dew"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/dns"
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/ges"
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/lts"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/nat"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/obs"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/sfs"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/swr"
 
-	dcs2 "github.com/sbercloud-terraform/terraform-provider-sbercloud/sbercloud/services/dcs"
-	dli2 "github.com/sbercloud-terraform/terraform-provider-sbercloud/sbercloud/services/dli"
+	elb2 "github.com/sbercloud-terraform/terraform-provider-sbercloud/sbercloud/services/elb"
 	"sync"
+
+	dcs2 "github.com/sbercloud-terraform/terraform-provider-sbercloud/sbercloud/services/dcs"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud"
@@ -45,6 +48,7 @@ import (
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/vpc"
 	"github.com/sbercloud-terraform/terraform-provider-sbercloud/sbercloud/services/ces"
 	"github.com/sbercloud-terraform/terraform-provider-sbercloud/sbercloud/services/css"
+	dli_sbercloud "github.com/sbercloud-terraform/terraform-provider-sbercloud/sbercloud/services/dli"
 	"github.com/sbercloud-terraform/terraform-provider-sbercloud/sbercloud/services/drs"
 )
 
@@ -201,7 +205,7 @@ func Provider() *schema.Provider {
 			"sbercloud_availability_zones":     huaweicloud.DataSourceAvailabilityZones(),
 			"sbercloud_css_flavors":            css_huawei.DataSourceCssFlavors(),
 			"sbercloud_cbr_backup":             cbr.DataSourceBackup(),
-			"sbercloud_cbr_vaults":             cbr.DataSourceCbrVaultsV3(),
+			"sbercloud_cbr_vaults":             cbr.DataSourceVaults(),
 			"sbercloud_cce_addon_template":     cce.DataSourceAddonTemplate(),
 			"sbercloud_cce_cluster":            cce.DataSourceCCEClusterV3(),
 			"sbercloud_cce_clusters":           cce.DataSourceCCEClusters(),
@@ -227,21 +231,21 @@ func Provider() *schema.Provider {
 			"sbercloud_elb_pools":              elb.DataSourcePools(),
 			"sbercloud_enterprise_project":     eps.DataSourceEnterpriseProject(),
 			"sbercloud_evs_volumes":            evs.DataSourceEvsVolumesV2(),
-			"sbercloud_identity_role":          iam.DataSourceIdentityRoleV3(),
+			"sbercloud_identity_role":          iam.DataSourceIdentityRole(),
 			"sbercloud_identity_custom_role":   iam.DataSourceIdentityCustomRole(),
 			"sbercloud_identity_group":         iam.DataSourceIdentityGroup(),
 			"sbercloud_identity_projects":      iam.DataSourceIdentityProjects(),
 			"sbercloud_identity_users":         iam.DataSourceIdentityUsers(),
 			"sbercloud_images_image":           ims.DataSourceImagesImageV2(),
 			"sbercloud_kms_key":                dew.DataSourceKmsKey(),
-			"sbercloud_kms_data_key":           huaweicloud.DataSourceKmsDataKeyV1(),
+			"sbercloud_kms_data_key":           dew.DataSourceKmsDataKeyV1(),
 			"sbercloud_lb_listeners":           lb.DataSourceListeners(),
 			"sbercloud_lb_loadbalancer":        lb.DataSourceELBV2Loadbalancer(),
 			"sbercloud_lb_certificate":         lb.DataSourceLBCertificateV2(),
 			"sbercloud_lb_pools":               lb.DataSourcePools(),
 			"sbercloud_nat_gateway":            nat.DataSourcePublicGateway(),
 			"sbercloud_networking_port":        vpc.DataSourceNetworkingPortV2(),
-			"sbercloud_networking_secgroup":    huaweicloud.DataSourceNetworkingSecGroup(),
+			"sbercloud_networking_secgroup":    vpc.DataSourceNetworkingSecGroup(),
 			"sbercloud_networking_secgroups":   vpc.DataSourceNetworkingSecGroups(),
 			"sbercloud_obs_buckets":            obs.DataSourceObsBuckets(),
 			"sbercloud_obs_bucket_object":      obs.DataSourceObsBucketObject(),
@@ -250,7 +254,7 @@ func Provider() *schema.Provider {
 			"sbercloud_rds_engine_versions":    rds.DataSourceRdsEngineVersionsV3(),
 			"sbercloud_rds_instances":          rds.DataSourceRdsInstances(),
 			"sbercloud_rds_storage_types":      rds.DataSourceStoragetype(),
-			"sbercloud_sfs_file_system":        huaweicloud.DataSourceSFSFileSystemV2(),
+			"sbercloud_sfs_file_system":        sfs.DataSourceSFSFileSystemV2(),
 			"sbercloud_sfs_turbos":             sfs.DataSourceTurbos(),
 			"sbercloud_vpc":                    vpc.DataSourceVpcV1(),
 			"sbercloud_vpcs":                   vpc.DataSourceVpcs(),
@@ -265,7 +269,7 @@ func Provider() *schema.Provider {
 			"sbercloud_vpc_subnets":            vpc.DataSourceVpcSubnets(),
 			"sbercloud_vpc_subnet_ids":         vpc.DataSourceVpcSubnetIdsV1(),
 			// Legacy
-			"sbercloud_identity_role_v3": iam.DataSourceIdentityRoleV3(),
+			"sbercloud_identity_role_v3": iam.DataSourceIdentityRole(),
 		},
 
 		ResourcesMap: map[string]*schema.Resource{
@@ -275,15 +279,15 @@ func Provider() *schema.Provider {
 			"sbercloud_as_configuration":                as.ResourceASConfiguration(),
 			"sbercloud_as_group":                        as.ResourceASGroup(),
 			"sbercloud_as_policy":                       as.ResourceASPolicy(),
-			"sbercloud_cbr_policy":                      cbr.ResourceCBRPolicyV3(),
+			"sbercloud_cbr_policy":                      cbr.ResourcePolicy(),
 			"sbercloud_cbr_vault":                       cbr.ResourceVault(),
 			"sbercloud_css_cluster":                     css.ResourceCssCluster(),
 			"sbercloud_cce_addon":                       cce.ResourceAddon(),
 			"sbercloud_cce_cluster":                     cce.ResourceCluster(),
 			"sbercloud_cce_namespace":                   cce.ResourceCCENamespaceV1(),
 			"sbercloud_cce_node":                        cce.ResourceNode(),
-			"sbercloud_cce_node_attach":                 cce.ResourceCCENodeAttachV3(),
-			"sbercloud_cce_node_pool":                   cce.ResourceCCENodePool(),
+			"sbercloud_cce_node_attach":                 cce.ResourceNodeAttach(),
+			"sbercloud_cce_node_pool":                   cce.ResourceNodePool(),
 			"sbercloud_cce_pvc":                         cce.ResourceCcePersistentVolumeClaimsV1(),
 			"sbercloud_cdm_cluster":                     cdm.ResourceCdmCluster(),
 			"sbercloud_compute_instance":                ResourceComputeInstanceV2(),
@@ -305,7 +309,7 @@ func Provider() *schema.Provider {
 			"sbercloud_dli_database":                    dli.ResourceDliSqlDatabaseV1(),
 			"sbercloud_dli_package":                     dli.ResourceDliPackageV2(),
 			"sbercloud_dli_queue":                       dli.ResourceDliQueue(),
-			"sbercloud_dli_spark_job":                   dli2.ResourceDliSparkJob(),
+			"sbercloud_dli_spark_job":                   dli_sbercloud.ResourceDliSparkJobV2(),
 			"sbercloud_dms_instance":                    ResourceDmsInstancesV1(),
 			"sbercloud_dms_kafka_instance":              dms.ResourceDmsKafkaInstance(),
 			"sbercloud_dms_kafka_topic":                 dms.ResourceDmsKafkaTopic(),
@@ -319,27 +323,27 @@ func Provider() *schema.Provider {
 			"sbercloud_elb_l7rule":                      elb.ResourceL7RuleV3(),
 			"sbercloud_elb_listener":                    elb.ResourceListenerV3(),
 			"sbercloud_elb_loadbalancer":                elb.ResourceLoadBalancerV3(),
-			"sbercloud_elb_monitor":                     elb.ResourceMonitorV3(),
+			"sbercloud_elb_monitor":                     elb2.ResourceMonitorV3(),
 			"sbercloud_elb_ipgroup":                     elb.ResourceIpGroupV3(),
 			"sbercloud_elb_pool":                        elb.ResourcePoolV3(),
 			"sbercloud_elb_member":                      elb.ResourceMemberV3(),
 			"sbercloud_elb_security_policy":             elb.ResourceSecurityPolicy(),
 			"sbercloud_enterprise_project":              eps.ResourceEnterpriseProject(),
-			"sbercloud_evs_snapshot":                    huaweicloud.ResourceEvsSnapshotV2(),
+			"sbercloud_evs_snapshot":                    evs.ResourceEvsSnapshotV2(),
 			"sbercloud_evs_volume":                      evs.ResourceEvsVolume(),
 			"sbercloud_fgs_function":                    fgs.ResourceFgsFunctionV2(),
-			"sbercloud_ges_graph":                       huaweicloud.ResourceGesGraphV1(),
+			"sbercloud_ges_graph":                       ges.ResourceGesGraph(),
 			"sbercloud_identity_access_key":             iam.ResourceIdentityKey(),
 			"sbercloud_identity_acl":                    iam.ResourceIdentityACL(),
 			"sbercloud_identity_agency":                 iam.ResourceIAMAgencyV3(),
-			"sbercloud_identity_group":                  iam.ResourceIdentityGroupV3(),
-			"sbercloud_identity_group_membership":       iam.ResourceIdentityGroupMembershipV3(),
-			"sbercloud_identity_project":                iam.ResourceIdentityProjectV3(),
+			"sbercloud_identity_group":                  iam.ResourceIdentityGroup(),
+			"sbercloud_identity_group_membership":       iam.ResourceIdentityGroupMembership(),
+			"sbercloud_identity_project":                iam.ResourceIdentityProject(),
 			"sbercloud_identity_provider":               iam.ResourceIdentityProvider(),
 			"sbercloud_identity_provider_conversion":    iam.ResourceIAMProviderConversion(),
 			"sbercloud_identity_role":                   iam.ResourceIdentityRole(),
 			"sbercloud_identity_role_assignment":        iam.ResourceIdentityGroupRoleAssignment(),
-			"sbercloud_identity_user":                   iam.ResourceIdentityUserV3(),
+			"sbercloud_identity_user":                   iam.ResourceIdentityUser(),
 			"sbercloud_images_image":                    ims.ResourceImsImage(),
 			"sbercloud_kms_key":                         dew.ResourceKmsKey(),
 			"sbercloud_lb_certificate":                  lb.ResourceCertificateV2(),
@@ -351,8 +355,8 @@ func Provider() *schema.Provider {
 			"sbercloud_lb_monitor":                      lb.ResourceMonitorV2(),
 			"sbercloud_lb_pool":                         lb.ResourcePoolV2(),
 			"sbercloud_lb_whitelist":                    lb.ResourceWhitelistV2(),
-			"sbercloud_lts_group":                       huaweicloud.ResourceLTSGroupV2(),
-			"sbercloud_lts_stream":                      huaweicloud.ResourceLTSStreamV2(),
+			"sbercloud_lts_group":                       lts.ResourceLTSGroup(),
+			"sbercloud_lts_stream":                      lts.ResourceLTSStream(),
 			"sbercloud_mapreduce_cluster":               mrs.ResourceMRSClusterV2(),
 			"sbercloud_mapreduce_job":                   mrs.ResourceMRSJobV2(),
 			"sbercloud_nat_dnat_rule":                   nat.ResourcePublicDnatRule(),
@@ -361,8 +365,8 @@ func Provider() *schema.Provider {
 			"sbercloud_network_acl":                     huaweicloud.ResourceNetworkACL(),
 			"sbercloud_network_acl_rule":                huaweicloud.ResourceNetworkACLRule(),
 			"sbercloud_networking_eip_associate":        eip.ResourceEIPAssociate(),
-			"sbercloud_networking_secgroup":             huaweicloud.ResourceNetworkingSecGroup(),
-			"sbercloud_networking_secgroup_rule":        huaweicloud.ResourceNetworkingSecGroupRule(),
+			"sbercloud_networking_secgroup":             vpc.ResourceNetworkingSecGroup(),
+			"sbercloud_networking_secgroup_rule":        vpc.ResourceNetworkingSecGroupRule(),
 			"sbercloud_networking_vip":                  vpc.ResourceNetworkingVip(),
 			"sbercloud_networking_vip_associate":        vpc.ResourceNetworkingVIPAssociateV2(),
 			"sbercloud_obs_bucket":                      obs.ResourceObsBucket(),
@@ -373,9 +377,9 @@ func Provider() *schema.Provider {
 			"sbercloud_rds_parametergroup":              rds.ResourceRdsConfiguration(),
 			"sbercloud_rds_backup":                      rds.ResourceBackup(),
 			"sbercloud_rds_read_replica_instance":       rds.ResourceRdsReadReplicaInstance(),
-			"sbercloud_sfs_access_rule":                 huaweicloud.ResourceSFSAccessRuleV2(),
-			"sbercloud_sfs_file_system":                 huaweicloud.ResourceSFSFileSystemV2(),
-			"sbercloud_sfs_turbo":                       huaweicloud.ResourceSFSTurbo(),
+			"sbercloud_sfs_access_rule":                 sfs.ResourceSFSAccessRuleV2(),
+			"sbercloud_sfs_file_system":                 sfs.ResourceSFSFileSystemV2(),
+			"sbercloud_sfs_turbo":                       sfs.ResourceSFSTurbo(),
 			"sbercloud_smn_subscription":                smn.ResourceSubscription(),
 			"sbercloud_smn_topic":                       smn.ResourceTopic(),
 			"sbercloud_swr_organization":                swr.ResourceSWROrganization(),
@@ -392,9 +396,9 @@ func Provider() *schema.Provider {
 			"sbercloud_vpc_address_group":               vpc.ResourceVpcAddressGroup(),
 			// Legacy
 			"sbercloud_identity_role_assignment_v3":  iam.ResourceIdentityGroupRoleAssignment(),
-			"sbercloud_identity_user_v3":             iam.ResourceIdentityUserV3(),
-			"sbercloud_identity_group_v3":            iam.ResourceIdentityGroupV3(),
-			"sbercloud_identity_group_membership_v3": iam.ResourceIdentityGroupMembershipV3(),
+			"sbercloud_identity_user_v3":             iam.ResourceIdentityUser(),
+			"sbercloud_identity_group_v3":            iam.ResourceIdentityGroup(),
+			"sbercloud_identity_group_membership_v3": iam.ResourceIdentityGroupMembership(),
 		},
 	}