+++ sbercloud/services/dli/resource_sbercloud_dli_package.go
@@ -1,0 +1,7 @@
+package dli
+
+const (
+	jarFile    = "jar"
+	pythonFile = "pyFile"
+	userFile   = "file"
+)