+++ sbercloud/services/dli/resource_sbercloud_dli_spark_job.go
@@ -18,13 +18,7 @@ import (
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils"
 )
 
-const (
-	jarFile    = "jar"
-	pythonFile = "pyFile"
-	userFile   = "file"
-)
-
-func ResourceDliSparkJob() *schema.Resource {
+func ResourceDliSparkJobV2() *schema.Resource {
 	return &schema.Resource{
 		CreateContext: ResourceDliSparkJobV2Create,
 		ReadContext:   ResourceDliSparkJobV2Read,